add kubectl-connect plugin
